+++ src/components/Liquidity/ExistingLiquidityPositions.tsx
@@ -4,9 +4,9 @@ import { useHistory } from 'react-router-dom'
 import AppTile from '../library/AppTile'
 import { useAppDispatch, useAppSelector as selector } from '../../app/hooks'
 import waitFor from '../../slices/waitFor'
-import { LiquidityPosition } from '../../slices/liquidityPositions'
+import { LiquidityPosition } from '../../slices/fetchNodes'
 import { bnf, tickToPriceDisplay, unscale, numDisplay, displaySymbol } from '../../utils/'
-import { poolMetadata } from '../../slices/poolsMetadata'
+import { poolMetadata } from '../../slices/fetchNodes'
 import Center from '../library/Center'
 import SimpleTable from '../library/SimpleTable'
 import SpacedList from '../library/SpacedList'
