+++ src/slices/contracts/ProtocolContract.ts
@@ -1,0 +1,26 @@
+enum ProtocolContract {
+  Accounting = "Accounting",
+  Auctions = "Auctions",
+  EnforcedDecentralization = "EnforcedDecentralization",
+  Hue = "Hue",
+  HuePositionNFT = "HuePositionNFT",
+  LendHue = "LendHue",
+  Liquidations = "Liquidations",
+  Market = "Market",
+  Prices = "Prices",
+  ProtocolLock = "ProtocolLock",
+  Rates = "Rates",
+  Rewards = "Rewards",
+  Settlement = "Settlement",
+  Tcp = "Tcp",
+  TcpGovernorAlpha = "TcpGovernorAlpha",
+  TcpTimelock = "TcpTimelock",
+}
+
+export enum RootContract {
+  Governor = "Governor",
+  ProtocolDataAggregator = "ProtocolDataAggregator",
+  TrustlessMulticall = "TrustlessMulticall",
+}
+
+export default ProtocolContract
