+++ src/slices/currentChainInfo/index.ts
@@ -1,5 +1,5 @@
 import { RootState } from '../../app/store'
-import { getThunkDependencies, NonNull } from '../waitFor'
+import { getThunkDependencies, NonNull, FetchNodes } from '../fetchNodes'
 import { sliceState } from '../'
 import { initialState, getGenericReducerBuilder } from '../'
 import { getMulticallContract} from '../../utils/getContract'
@@ -10,12 +10,6 @@ import {
 } from '@trustlessfi/multicall'
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
 
-export interface currentChainInfo {
-  blockNumber: number
-  blockTimestamp: number
-  chainID: number
-}
-
 const dependencies = getThunkDependencies(['trustlessMulticall'])
 
 export const getCurrentChainInfo = {
@@ -24,7 +18,7 @@ export const getCurrentChainInfo = {
   thunk:
     createAsyncThunk(
       'currentChainInfo/getCurrentChainInfo',
-      async (args: NonNull<typeof dependencies>): Promise<currentChainInfo> => {
+      async (args: NonNull<typeof dependencies>) => {
         const multicall = getMulticallContract(args.trustlessMulticall)
 
         const { chainInfo } = await executeMulticalls(
@@ -52,7 +46,7 @@ export const getCurrentChainInfo = {
 
 export const currentChainInfoSlice = createSlice({
   name: 'currentChainInfo',
-  initialState: initialState as sliceState<currentChainInfo>,
+  initialState: initialState as sliceState<FetchNodes['currentChainInfo']>,
   reducers: {},
   extraReducers: (builder) => {
     builder = getGenericReducerBuilder(builder, getCurrentChainInfo.thunk)
