+++ src/slices/liquidations/index.ts
@@ -1,19 +1,13 @@
 import { RootState } from '../../app/store'
-import { getThunkDependencies, NonNull } from '../waitFor'
+import { getThunkDependencies, NonNull, FetchNodes } from '../fetchNodes'
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
 import { sliceState, getStateWithValue, getGenericReducerBuilder } from '../'
 import getContract, { getMulticallContract } from '../../utils/getContract'
 
 import { Liquidations } from '@trustlessfi/typechain'
-import { ProtocolContract } from '../contracts'
+import ProtocolContract from '../contracts/ProtocolContract'
 import { getLocalStorage } from '../../utils'
 import { executeMulticalls, oneContractManyFunctionMC, rc } from '@trustlessfi/multicall'
-
-export type liquidationsInfo = {
-  twapDuration: number,
-  discoveryIncentive: number,
-  liquidationIncentive: number,
-}
 
 const dependencies = getThunkDependencies(['contracts', 'trustlessMulticall'])
 
@@ -23,7 +17,7 @@ export const getLiquidationsInfo = {
   thunk:
     createAsyncThunk(
     'liquidations/getLiquidationsInfo',
-    async (args: NonNull<typeof dependencies>): Promise<liquidationsInfo> => {
+    async (args: NonNull<typeof dependencies>) => {
       const liquidations = getContract(args.contracts[ProtocolContract.Liquidations], ProtocolContract.Liquidations) as Liquidations
       const trustlessMulticall = getMulticallContract(args.trustlessMulticall)
 
@@ -50,7 +44,7 @@ const name = 'liquidations'
 
 export const liquidationsSlice = createSlice({
   name,
-  initialState: getStateWithValue(getLocalStorage(name)) as sliceState<liquidationsInfo>,
+  initialState: getStateWithValue(getLocalStorage(name)) as sliceState<FetchNodes['liquidationsInfo']>,
   reducers: {},
   extraReducers: (builder) => {
     builder = getGenericReducerBuilder(builder, getLiquidationsInfo.thunk)
