+++ src/slices/poolsCurrentData/index.ts
@@ -1,11 +1,12 @@
 import { RootState } from '../../app/store'
+import { getThunkDependencies, NonNull, FetchNodes } from '../fetchNodes'
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { initialState, getGenericReducerBuilder } from '../'
 import { sliceState } from '../'
 
 import { Contract } from 'ethers'
 
-import { ProtocolContract } from '../contracts'
+import ProtocolContract from '../contracts/ProtocolContract'
 import getProvider from '../../utils/getProvider'
 import getContract, { getMulticallContract } from '../../utils/getContract'
 import {
@@ -17,28 +18,11 @@ import {
   idToIdAndNoArg,
   idToIdAndArg,
 } from '@trustlessfi/multicall'
-import { NonNull, getThunkDependencies } from '../waitFor'
 
 import { Prices, UniswapV3Pool, Accounting, Rewards } from '@trustlessfi/typechain'
 
 import poolArtifact from '@trustlessfi/artifacts/dist/@uniswap/v3-core/contracts/interfaces/IUniswapV3Pool.sol/IUniswapV3Pool.json'
 import { sqrtPriceX96ToTick, zeroAddress, PromiseType } from '../../utils'
-
-export type poolCurrentInfo = {
-  instantTick: number
-  twapTick: number
-  poolLiquidity: string
-  cumulativeLiquidity: string
-  totalRewards: string
-  lastPeriodGlobalRewardsAccrued: number
-  currentPeriod: number
-}
-
-export type poolsCurrentInfo = {
-  [key in string]: poolCurrentInfo
-}
-
-export type poolCurrentDataState = sliceState<poolsCurrentInfo>
 
 
 const dependencies = getThunkDependencies(['contracts', 'trustlessMulticall', 'poolsMetadata', 'rewardsInfo'])
@@ -48,7 +32,7 @@ export const getPoolsCurrentData = {
   dependencies,
   thunk: createAsyncThunk(
     'poolsCurrentData/getPoolsCurrentData',
-    async (args: NonNull<typeof dependencies>): Promise<poolsCurrentInfo> => {
+    async (args: NonNull<typeof dependencies>) => {
       const provider = getProvider()
       const prices = getContract(args.contracts[ProtocolContract.Prices], ProtocolContract.Prices) as Prices
       const rewards = getContract(args.contracts[ProtocolContract.Rewards], ProtocolContract.Rewards) as Rewards
@@ -116,7 +100,7 @@ export const getPoolsCurrentData = {
 
 export const poolsCurrentDataSlice = createSlice({
   name: 'poolsCurrentData' ,
-  initialState: initialState as poolCurrentDataState,
+  initialState: initialState as sliceState<FetchNodes['poolsCurrentData']>,
   reducers: {
     clearPoolsCurrentData: (state) => {
       state.value = null
