+++ src/slices/rates/index.ts
@@ -1,19 +1,13 @@
 import { RootState } from '../../app/store'
-import { getThunkDependencies, NonNull } from '../waitFor'
+import { getThunkDependencies, NonNull, FetchNodes } from '../fetchNodes'
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { sliceState, getStateWithValue, getGenericReducerBuilder } from '../'
 import getContract, { getMulticallContract } from '../../utils/getContract'
 
 import { Rates } from '@trustlessfi/typechain/'
-import { ProtocolContract } from '../contracts'
+import ProtocolContract from '../contracts/ProtocolContract'
 import { getLocalStorage } from '../../utils'
 import { oneContractManyFunctionMC, rc, executeMulticalls } from '@trustlessfi/multicall'
-
-export type ratesInfo = {
-  positiveInterestRate: boolean,
-  interestRateAbsoluteValue: number,
-  referencePools: string[]
-}
 
 const dependencies = getThunkDependencies(['contracts', 'trustlessMulticall'])
 
@@ -23,7 +17,7 @@ export const getRatesInfo = {
   thunk:
     createAsyncThunk(
       'rates/getRatesInfo',
-      async (args: NonNull<typeof dependencies>): Promise<ratesInfo> => {
+      async (args: NonNull<typeof dependencies>) => {
         const rates = getContract(args.contracts[ProtocolContract.Rates], ProtocolContract.Rates) as Rates
         const trustlessMulticall = getMulticallContract(args.trustlessMulticall)
 
@@ -54,7 +48,7 @@ const name = 'rates'
 
 export const ratesSlice = createSlice({
   name,
-  initialState: getStateWithValue(getLocalStorage(name)) as sliceState<ratesInfo>,
+  initialState: getStateWithValue(getLocalStorage(name)) as sliceState<FetchNodes['ratesInfo']>,
   reducers: {},
   extraReducers: (builder) => {
     builder = getGenericReducerBuilder(builder, getRatesInfo.thunk)
