+++ src/slices/systemDebt/index.ts
@@ -1,20 +1,13 @@
 import { RootState } from '../../app/store'
-import { getThunkDependencies, NonNull } from '../waitFor'
+import { getThunkDependencies, NonNull, FetchNodes } from '../fetchNodes'
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { sliceState, getStateWithValue, getGenericReducerBuilder } from '../'
 import { unscale } from '../../utils'
 import getContract from '../../utils/getContract'
 
 import { Accounting } from '@trustlessfi/typechain';
-import { ProtocolContract, contractsInfo } from '../contracts'
+import ProtocolContract from '../contracts/ProtocolContract'
 import { getLocalStorage } from '../../utils'
-
-export type systemDebtInfo = {
-  debt: number
-  totalTCPRewards: number
-  cumulativeDebt: number
-  debtExchangeRate: number
-}
 
 const dependencies = getThunkDependencies(['contracts'])
 
@@ -43,7 +36,7 @@ const name = 'systemDebt'
 
 export const systemDebtSlice = createSlice({
   name,
-  initialState: getStateWithValue<systemDebtInfo>(getLocalStorage(name, null)) as sliceState<systemDebtInfo>,
+  initialState: getStateWithValue(getLocalStorage(name)) as sliceState<FetchNodes['sdi']>,
   reducers: {},
   extraReducers: (builder) => {
     builder = getGenericReducerBuilder(builder, getSystemDebtInfo.thunk)
