+++ src/slices/waitFor.ts
@@ -1,60 +1,27 @@
 import { AsyncThunk } from '@reduxjs/toolkit'
 import { AppDispatch, store, RootState } from '../app/store'
-import { ChainID } from '@trustlessfi/addresses'
 import { AppSelector } from '../app/hooks'
-import { getGovernorInfo, governorInfo } from './governor'
-import { getMarketInfo, marketInfo } from './market'
-import { getRatesInfo, ratesInfo } from './rates'
-import { getPoolsMetadata, poolsMetadata } from './poolsMetadata'
-import { getPoolsCurrentData, poolsCurrentInfo } from './poolsCurrentData'
-import { getLiquidityPositions, liquidityPositions } from './liquidityPositions'
-import { getPositions, positionsInfo } from './positions'
-import { getSystemDebtInfo, systemDebtInfo } from './systemDebt'
-import { getLiquidationsInfo, liquidationsInfo } from './liquidations'
-import { getRewardsInfo, rewardsInfo } from './rewards'
-import { getPricesInfo, pricesInfo } from './prices'
-import { getBalances, balancesInfo } from './balances'
-import { getContracts, contractsInfo } from './contracts'
-import { getCurrentChainInfo, currentChainInfo } from './currentChainInfo'
+import { getGovernorInfo } from './governor'
+import { getMarketInfo } from './market'
+import { getRatesInfo } from './rates'
+import { getPoolsMetadata } from './poolsMetadata'
+import { getPoolsCurrentData } from './poolsCurrentData'
+import { getLiquidityPositions } from './liquidityPositions'
+import { getPositions } from './positions'
+import { getSystemDebtInfo } from './systemDebt'
+import { getLiquidationsInfo } from './liquidations'
+import { getRewardsInfo } from './rewards'
+import { getPricesInfo } from './prices'
+import { getBalances } from './balances'
+import { getContracts } from './contracts'
+import { getCurrentChainInfo } from './currentChainInfo'
 
 import { sliceState } from './'
-
-const toNull = <T>() => null as T | null
-const nullString = toNull<string>()
-
-const fetchNodes = {
-  chainID: toNull<ChainID>(),
-  governor: nullString,
-  protocolDataAggregator: nullString,
-  trustlessMulticall: nullString,
-  userAddress: nullString,
-
-  balances: toNull<balancesInfo>(),
-  contracts: toNull<contractsInfo>(),
-  currentChainInfo: toNull<currentChainInfo>(),
-  governorInfo: toNull<governorInfo>(),
-  liquidationsInfo: toNull<liquidationsInfo>(),
-  liquidityPositions: toNull<liquidityPositions>(),
-  marketInfo: toNull<marketInfo>(),
-  poolsCurrentData: toNull<poolsCurrentInfo>(),
-  poolsMetadata: toNull<poolsMetadata>(),
-  positions: toNull<positionsInfo>(),
-  pricesInfo: toNull<pricesInfo>(),
-  ratesInfo: toNull<ratesInfo>(),
-  rewardsInfo: toNull<rewardsInfo>(),
-  sdi: toNull<systemDebtInfo>(),
-}
-
-export type NonNull<O> = {
-  [K in keyof O]-?: NonNullable<O[K]>
-}
-
-type FetchNodes = typeof fetchNodes
-type FetchNode = keyof FetchNodes
+import { NonNull, FetchNode, FetchNodes } from './fetchNodes'
 
 const getWaitFunction = <
     Value,
-    Dependencies extends Partial<{[fetchNode in keyof typeof fetchNodes]: (typeof fetchNodes)[fetchNode] | null}>,
+    Dependencies extends Partial<{[fetchNode in keyof FetchNodes]: FetchNodes[fetchNode] | null}>,
   >(waitForData: {
     stateSelector: (state: RootState) => sliceState<Value>
     dependencies: Dependencies
@@ -85,20 +52,10 @@ const getWaitFunction = <
     return state === undefined ? null : state.value
   }
 
-const allPossibleArgs =
-  Object.fromEntries(
-    Object.entries(fetchNodes)
-      .map(([key, _val]) => [key, null])
-  ) as { [key in FetchNode]: FetchNodes[key] }
-
-export const getThunkDependencies = <R extends FetchNode>(val: R[]) =>
-  Object.fromEntries(val.map(key => [key, allPossibleArgs[key]])) as Pick<typeof allPossibleArgs, R>
-
-
 const getStateSelector = <T>(selectorFunc: (state: RootState) => T) =>
   (selector: AppSelector, _dispatch: AppDispatch) => selector(selectorFunc)
 
-const waitForImpl: {[key in FetchNode]: (selector: AppSelector, _dispatch: AppDispatch) => FetchNodes[key]} = {
+const waitForImpl: {[key in FetchNode]: (selector: AppSelector, _dispatch: AppDispatch) => FetchNodes[key] | null} = {
   chainID: getStateSelector(state => state.chainID.chainID),
   governor: getStateSelector(state => state.chainID.governor),
   protocolDataAggregator: getStateSelector(state => state.chainID.protocolDataAggregator),
